+++ src/LiveComponent/assets/src/clone_html_element.ts
@@ -1,0 +1,8 @@
+export function cloneHTMLElement(element: HTMLElement): HTMLElement {
+    const newElement = element.cloneNode(true);
+    if (!(newElement instanceof HTMLElement)) {
+        throw new Error('Could not clone element');
+    }
+
+    return newElement;
+}